backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
gunicorn==21.2.0
Pillow==10.3.0
